requirements.txt
# requirements_new.txt - For Python 3.13
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.0
tldextract==5.1.1
pyyaml==6.0
requests==2.31.0